+++ platform/ios/src/MGLMapView.mm
@@ -3526,6 +3526,28 @@ public:
     return *self.mbglMap.getBounds().maxZoom;
 }
 
+- (CGFloat)minimumPitch
+{
+    return *_mbglMap->getBounds().minPitch;
+}
+
+- (void)setMinimumPitch:(CGFloat)minimumPitch
+{
+    MGLLogDebug(@"Setting minimumPitch: %f", minimumPitch);
+    _mbglMap->setBounds(mbgl::BoundOptions().withMinPitch(minimumPitch));
+}
+
+- (CGFloat)maximumPitch
+{
+    return *_mbglMap->getBounds().maxPitch;
+}
+
+- (void)setMaximumPitch:(CGFloat)maximumPitch
+{
+    MGLLogDebug(@"Setting maximumPitch: %f", maximumPitch);
+    _mbglMap->setBounds(mbgl::BoundOptions().withMaxPitch(maximumPitch));
+}
+
 - (MGLCoordinateBounds)visibleCoordinateBounds
 {
     return [self convertRect:self.bounds toCoordinateBoundsFromView:self];